Move shims into interchange::cjson::shims
It's possible that future interchange formats will have a different
metadata layout than the cjson format. Since our current shims are
then intrinsically tied to cjson, the shims themselvs should be in
the interchange module.
