+++ src/metadata.rs
@@ -14,7 +14,7 @@ use std::marker::PhantomData;
 use crate::crypto::{self, HashAlgorithm, HashValue, KeyId, PrivateKey, PublicKey, Signature};
 use crate::error::Error;
 use crate::interchange::DataInterchange;
-use crate::shims;
+use crate::interchange::cjson::shims;
 use crate::Result;
 
 #[rustfmt::skip]
